Int to float and step 2 copy temperature fix

+++ src/UserDashboard/pages/HeatingSchedule/CloneHeating/CloneHeatingModal.jsx
@@ -441,7 +441,7 @@ export function CloneHeatingModal({
 						entries.forEach((entry) => {
 							const from = normalizeTime(entry.y);
 							let to = normalizeTime(entry.y + entry.h);
-							const targetTemperature = parseInt(entry.temperature, 10);
+							const targetTemperature = parseFloat(entry.temperature, 10);
 
 							if (to === "24:00") {
 								to = "23:59";
@@ -472,10 +472,10 @@ export function CloneHeatingModal({
 					days: convertScheduleData(combinedData.finalScheduleData),
 				};
 				if (combinedData.formData.childSafety !== "Yes") {
-					finalObj.deviceOverrideTemperatureMin = parseInt(
+					finalObj.deviceOverrideTemperatureMin = parseFloat(
 						combinedData.formData.minTemp,
 					);
-					finalObj.deviceOverrideTemperatureMax = parseInt(
+					finalObj.deviceOverrideTemperatureMax = parseFloat(
 						combinedData.formData.maxTemp,
 					);
 				}
