+++ src/UserDashboard/pages/HeatingSchedule/EditHeating/EditHeatingModal.jsx
@@ -385,7 +385,7 @@ export function EditHeatingModal({
 						entries.forEach((entry) => {
 							const from = normalizeTime(entry.y);
 							let to = normalizeTime(entry.y + entry.h);
-							const targetTemperature = parseInt(entry.temperature, 10);
+							const targetTemperature = parseFloat(entry.temperature, 10);
 
 							if (to === "24:00") {
 								to = "23:59";
@@ -412,10 +412,10 @@ export function EditHeatingModal({
 				};
 
 				if (combinedData.formData.childSafety !== "Yes") {
-					finalObj.deviceOverrideTemperatureMin = parseInt(
+					finalObj.deviceOverrideTemperatureMin = parseFloat(
 						combinedData.formData.minTemp,
 					);
-					finalObj.deviceOverrideTemperatureMax = parseInt(
+					finalObj.deviceOverrideTemperatureMax = parseFloat(
 						combinedData.formData.maxTemp,
 					);
 				}
